Restructure files to break down constants into smaller parts for readability/start adding intermediate classes so data chance underneath doesn't affect API calls
