+++ pyboy/plugins/game_wrapper_pokemon_gen1/constants.py
@@ -169,198 +169,6 @@ class Moves(Enum):
     SLASH = 0xa3
     SUBSTITUTE = 0xa4
 
-class Pokemon(Enum):
-    RHYDON = 0x01
-    KANGASKHAN = 0x02
-    NIDORAN_M = 0x03
-    CLEFAIRY = 0x04
-    SPEAROW = 0x05
-    VOLTORB = 0x06
-    NIDOKING = 0x07
-    SLOWBRO = 0x08
-    IVYSAUR = 0x09
-    EXEGGUTOR = 0x0A
-    LICKITUNG = 0x0B
-    EXEGGCUTE = 0x0C
-    GRIMER = 0x0D
-    GENGAR = 0x0E
-    NIDORAN_F = 0x0F
-    NIDOQUEEN = 0x10
-    CUBONE = 0x11
-    RHYHORN = 0x12
-    LAPRAS = 0x13
-    ARCANINE = 0x14
-    MEW = 0x15
-    GYARADOS = 0x16
-    SHELLDER = 0x17
-    TENTACOOL = 0x18
-    GASTLY = 0x19
-    SCYTHER = 0x1A
-    STARYU = 0x1B
-    BLASTOISE = 0x1C
-    PINSIR = 0x1D
-    TANGELA = 0x1E
-    MISSINGNO_1F = 0x1F
-    MISSINGNO_20 = 0x20
-    GROWLITHE = 0x21
-    ONIX = 0x22
-    FEAROW = 0x23
-    PIDGEY = 0x24
-    SLOWPOKE = 0x25
-    KADABRA = 0x26
-    GRAVELER = 0x27
-    CHANSEY = 0x28
-    MACHOKE = 0x29
-    MR_MIME = 0x2A
-    HITMONLEE = 0x2B
-    HITMONCHAN = 0x2C
-    ARBOK = 0x2D
-    PARASECT = 0x2E
-    PSYDUCK = 0x2F
-    DROWZEE = 0x30
-    GOLEM = 0x31
-    MISSINGNO_32 = 0x32
-    MAGMAR = 0x33
-    MISSINGNO_34 = 0x34
-    ELECTABUZZ = 0x35
-    MAGNETON = 0x36
-    KOFFING = 0x37
-    MISSINGNO_38 = 0x38
-    MANKEY = 0x39
-    SEEL = 0x3A
-    DIGLETT = 0x3B
-    TAUROS = 0x3C
-    MISSINGNO_3D = 0x3D
-    MISSINGNO_3E = 0x3E
-    MISSINGNO_3F = 0x3F
-    FARFETCHD = 0x40
-    VENONAT = 0x41
-    DRAGONITE = 0x42
-    MISSINGNO_43 = 0x43
-    MISSINGNO_44 = 0x44
-    MISSINGNO_45 = 0x45
-    DODUO = 0x46
-    POLIWAG = 0x47
-    JYNX = 0x48
-    MOLTRES = 0x49
-    ARTICUNO = 0x4A
-    ZAPDOS = 0x4B
-    DITTO = 0x4C
-    MEOWTH = 0x4D
-    KRABBY = 0x4E
-    MISSINGNO_4F = 0x4F
-    MISSINGNO_50 = 0x50
-    MISSINGNO_51 = 0x51
-    VULPIX = 0x52
-    NINETALES = 0x53
-    PIKACHU = 0x54
-    RAICHU = 0x55
-    MISSINGNO_56 = 0x56
-    MISSINGNO_57 = 0x57
-    DRATINI = 0x58
-    DRAGONAIR = 0x59
-    KABUTO = 0x5A
-    KABUTOPS = 0x5B
-    HORSEA = 0x5C
-    SEADRA = 0x5D
-    MISSINGNO_5E = 0x5E
-    MISSINGNO_5F = 0x5F
-    SANDSHREW = 0x60
-    SANDSLASH = 0x61
-    OMANYTE = 0x62
-    OMASTAR = 0x63
-    JIGGLYPUFF = 0x64
-    WIGGLYTUFF = 0x65
-    EEVEE = 0x66
-    FLAREON = 0x67
-    JOLTEON = 0x68
-    VAPOREON = 0x69
-    MACHOP = 0x6A
-    ZUBAT = 0x6B
-    EKANS = 0x6C
-    PARAS = 0x6D
-    POLIWHIRL = 0x6E
-    POLIWRATH = 0x6F
-    WEEDLE = 0x70
-    KAKUNA = 0x71
-    BEEDRILL = 0x72
-    MISSINGNO_73 = 0x73
-    DODRIO = 0x74
-    PRIMEAPE = 0x75
-    DUGTRIO = 0x76
-    VENOMOTH = 0x77
-    DEWGONG = 0x78
-    MISSINGNO_79 = 0x79
-    MISSINGNO_7A = 0x7A
-    CATERPIE = 0x7B
-    METAPOD = 0x7C
-    BUTTERFREE = 0x7D
-    MACHAMP = 0x7E
-    MISSINGNO_7F = 0x7F
-    GOLDUCK = 0x80
-    HYPNO = 0x81
-    GOLBAT = 0x82
-    MEWTWO = 0x83
-    SNORLAX = 0x84
-    MAGIKARP = 0x85
-    MISSINGNO_86 = 0x86
-    MISSINGNO_87 = 0x87
-    MUK = 0x88
-    MISSINGNO_89 = 0x89
-    KINGLER = 0x8A
-    CLOYSTER = 0x8B
-    MISSINGNO_8C = 0x8C
-    ELECTRODE = 0x8D
-    CLEFABLE = 0x8E
-    WEEZING = 0x8F
-    PERSIAN = 0x90
-    MAROWAK = 0x91
-    MISSINGNO_92 = 0x92
-    HAUNTER = 0x93
-    ABRA = 0x94
-    ALAKAZAM = 0x95
-    PIDGEOTTO = 0x96
-    PIDGEOT = 0x97
-    STARMIE = 0x98
-    BULBASAUR = 0x99
-    VENUSAUR = 0x9A
-    TENTACRUEL = 0x9B
-    MISSINGNO_9C = 0x9C
-    GOLDEEN = 0x9D
-    SEAKING = 0x9E
-    MISSINGNO_9F = 0x9F
-    MISSINGNO_A0 = 0xA0
-    MISSINGNO_A1 = 0xA1
-    MISSINGNO_A2 = 0xA2
-    PONYTA = 0xA3
-    RAPIDASH = 0xA4
-    RATTATA = 0xA5
-    RATICATE = 0xA6
-    NIDORINO = 0xA7
-    NIDORINA = 0xA8
-    GEODUDE = 0xA9
-    PORYGON = 0xAA
-    AERODACTYL = 0xAB
-    MISSINGNO_AC = 0xAC
-    MAGNEMITE = 0xAD
-    MISSINGNO_AE = 0xAE
-    MISSINGNO_AF = 0xAF
-    CHARMANDER = 0xB0
-    SQUIRTLE = 0xB1
-    CHARMELEON = 0xB2
-    WARTORTLE = 0xB3
-    CHARIZARD = 0xB4
-    MISSINGNO_B5 = 0xB5
-    FOSSIL_KABUTOPS = 0xB6
-    FOSSIL_AERODACTYL = 0xB7
-    MON_GHOST = 0xB8
-    ODDISH = 0xB9
-    GLOOM = 0xBA
-    VILEPLUME = 0xBB
-    BELLSPROUT = 0xBC
-    WEEPINBELL = 0xBD
-    VICTREEBEL = 0xBE
-
 class Items(Enum):
     MASTER_BALL = 0x01
     ULTRA_BALL = 0x02
@@ -1096,466 +904,3 @@ class Statuses(Enum):
     BURNED = 16
     FROZEN = 32
     PARALYZED = 64
-
-NAMES = (
-    "Bulbasaur",
-    "Ivysaur",
-    "Venusaur",
-    "Charmander",
-    "Charmeleon",
-    "Charizard",
-    "Squirtle",
-    "Wartortle",
-    "Blastoise",
-    "Caterpie",
-    "Metapod",
-    "Butterfree",
-    "Weedle",
-    "Kakuna",
-    "Beedrill",
-    "Pidgey",
-    "Pidgeotto",
-    "Pidgeot",
-    "Rattata",
-    "Raticate",
-    "Spearow",
-    "Fearow",
-    "Ekans",
-    "Arbok",
-    "Pikachu",
-    "Raichu",
-    "Sandshrew",
-    "Sandslash",
-    "Nidoran♀",
-    "Nidorina",
-    "Nidoqueen",
-    "Nidoran♂",
-    "Nidorino",
-    "Nidoking",
-    "Clefairy",
-    "Clefable",
-    "Vulpix",
-    "Ninetales",
-    "Jigglypuff",
-    "Wigglytuff",
-    "Zubat",
-    "Golbat",
-    "Oddish",
-    "Gloom",
-    "Vileplume",
-    "Paras",
-    "Parasect",
-    "Venonat",
-    "Venomoth",
-    "Diglett",
-    "Dugtrio",
-    "Meowth",
-    "Persian",
-    "Psyduck",
-    "Golduck",
-    "Mankey",
-    "Primeape",
-    "Growlithe",
-    "Arcanine",
-    "Poliwag",
-    "Poliwhirl",
-    "Poliwrath",
-    "Abra",
-    "Kadabra",
-    "Alakazam",
-    "Machop",
-    "Machoke",
-    "Machamp",
-    "Bellsprout",
-    "Weepinbell",
-    "Victreebel",
-    "Tentacool",
-    "Tentacruel",
-    "Geodude",
-    "Graveler",
-    "Golem",
-    "Ponyta",
-    "Rapidash",
-    "Slowpoke",
-    "Slowbro",
-    "Magnemite",
-    "Magneton",
-    "Farfetch’d",
-    "Doduo",
-    "Dodrio",
-    "Seel",
-    "Dewgong",
-    "Grimer",
-    "Muk",
-    "Shellder",
-    "Cloyster",
-    "Gastly",
-    "Haunter",
-    "Gengar",
-    "Onix",
-    "Drowzee",
-    "Hypno",
-    "Krabby",
-    "Kingler",
-    "Voltorb",
-    "Electrode",
-    "Exeggcute",
-    "Exeggutor",
-    "Cubone",
-    "Marowak",
-    "Hitmonlee",
-    "Hitmonchan",
-    "Lickitung",
-    "Koffing",
-    "Weezing",
-    "Rhyhorn",
-    "Rhydon",
-    "Chansey",
-    "Tangela",
-    "Kangaskhan",
-    "Horsea",
-    "Seadra",
-    "Goldeen",
-    "Seaking",
-    "Staryu",
-    "Starmie",
-    "Mr. Mime",
-    "Scyther",
-    "Jynx",
-    "Electabuzz",
-    "Magmar",
-    "Pinsir",
-    "Tauros",
-    "Magikarp",
-    "Gyarados",
-    "Lapras",
-    "Ditto",
-    "Eevee",
-    "Vaporeon",
-    "Jolteon",
-    "Flareon",
-    "Porygon",
-    "Omanyte",
-    "Omastar",
-    "Kabuto",
-    "Kabutops",
-    "Aerodactyl",
-    "Snorlax",
-    "Articuno",
-    "Zapdos",
-    "Moltres",
-    "Dratini",
-    "Dragonair",
-    "Dragonite",
-    "Mewtwo",
-    "Mew",
-)
-
-POKEMON_POKEDEX_INDEX = {
-    Pokemon.BULBASAUR: 1,
-    Pokemon.IVYSAUR:  2,
-    Pokemon.VENUSAUR: 3,
-    Pokemon.CHARMANDER: 4,
-    Pokemon.CHARMELEON: 5,
-    Pokemon.CHARIZARD: 6,
-    Pokemon.SQUIRTLE: 7,
-    Pokemon.WARTORTLE: 8,
-    Pokemon.BLASTOISE: 9,
-    Pokemon.CATERPIE: 10,
-    Pokemon.METAPOD: 11,
-    Pokemon.BUTTERFREE: 12,
-    Pokemon.WEEDLE: 13,
-    Pokemon.KAKUNA: 14,
-    Pokemon.BEEDRILL: 15,
-    Pokemon.PIDGEY: 16,
-    Pokemon.PIDGEOTTO: 17,
-    Pokemon.PIDGEOT: 18,
-    Pokemon.RATTATA: 19,
-    Pokemon.RATICATE: 20,
-    Pokemon.SPEAROW: 21,
-    Pokemon.FEAROW: 22,
-    Pokemon.EKANS: 23,
-    Pokemon.ARBOK: 24,
-    Pokemon.PIKACHU: 25,
-    Pokemon.RAICHU: 26,
-    Pokemon.SANDSHREW: 27,
-    Pokemon.SANDSLASH: 28,
-    Pokemon.NIDORAN_F: 29,
-    Pokemon.NIDORINA: 30,
-    Pokemon.NIDOQUEEN: 31,
-    Pokemon.NIDORAN_M: 32,
-    Pokemon.NIDORINO: 33,
-    Pokemon.NIDOKING: 34,
-    Pokemon.CLEFAIRY: 35,
-    Pokemon.CLEFABLE: 36,
-    Pokemon.VULPIX: 37,
-    Pokemon.NINETALES: 38,
-    Pokemon.JIGGLYPUFF: 39,
-    Pokemon.WIGGLYTUFF: 40,
-    Pokemon.ZUBAT: 41,
-    Pokemon.GOLBAT: 42,
-    Pokemon.ODDISH: 43,
-    Pokemon.GLOOM: 44,
-    Pokemon.VILEPLUME: 45,
-    Pokemon.PARAS: 46,
-    Pokemon.PARASECT: 47,
-    Pokemon.VENONAT: 48,
-    Pokemon.VENOMOTH: 49,
-    Pokemon.DIGLETT: 50,
-    Pokemon.DUGTRIO: 51,
-    Pokemon.MEOWTH: 52,
-    Pokemon.PERSIAN: 53,
-    Pokemon.PSYDUCK: 54,
-    Pokemon.GOLDUCK: 55,
-    Pokemon.MANKEY: 56,
-    Pokemon.PRIMEAPE: 57,
-    Pokemon.GROWLITHE: 58,
-    Pokemon.ARCANINE: 59,
-    Pokemon.POLIWAG: 60,
-    Pokemon.POLIWHIRL: 61,
-    Pokemon.POLIWRATH: 62,
-    Pokemon.ABRA: 63,
-    Pokemon.KADABRA: 64,
-    Pokemon.ALAKAZAM: 65,
-    Pokemon.MACHOP: 66,
-    Pokemon.MACHOKE: 67,
-    Pokemon.MACHAMP: 68,
-    Pokemon.BELLSPROUT: 69,
-    Pokemon.WEEPINBELL: 70,
-    Pokemon.VICTREEBEL: 71,
-    Pokemon.TENTACOOL: 72,
-    Pokemon.TENTACRUEL: 73,
-    Pokemon.GEODUDE: 74,
-    Pokemon.GRAVELER: 75,
-    Pokemon.GOLEM: 76,
-    Pokemon.PONYTA: 77,
-    Pokemon.RAPIDASH: 78,
-    Pokemon.SLOWPOKE: 79,
-    Pokemon.SLOWBRO: 80,
-    Pokemon.MAGNEMITE: 81,
-    Pokemon.MAGNETON: 82,
-    Pokemon.FARFETCHD: 83,
-    Pokemon.DODUO: 84,
-    Pokemon.DODRIO: 85,
-    Pokemon.SEEL: 86,
-    Pokemon.DEWGONG: 87,
-    Pokemon.GRIMER: 88,
-    Pokemon.MUK: 89,
-    Pokemon.SHELLDER: 90,
-    Pokemon.CLOYSTER: 91,
-    Pokemon.GASTLY: 92,
-    Pokemon.HAUNTER: 93,
-    Pokemon.GENGAR: 94,
-    Pokemon.ONIX: 95,
-    Pokemon.DROWZEE: 96,
-    Pokemon.HYPNO: 97,
-    Pokemon.KRABBY: 98,
-    Pokemon.KINGLER: 99,
-    Pokemon.VOLTORB: 100,
-    Pokemon.ELECTRODE: 101,
-    Pokemon.EXEGGCUTE: 102,
-    Pokemon.EXEGGUTOR: 103,
-    Pokemon.CUBONE: 104,
-    Pokemon.MAROWAK: 105,
-    Pokemon.HITMONLEE: 106,
-    Pokemon.HITMONCHAN: 107,
-    Pokemon.LICKITUNG: 108,
-    Pokemon.KOFFING: 109,
-    Pokemon.WEEZING: 110,
-    Pokemon.RHYHORN: 111,
-    Pokemon.RHYDON: 112,
-    Pokemon.CHANSEY: 113,
-    Pokemon.TANGELA: 114,
-    Pokemon.KANGASKHAN: 115,
-    Pokemon.HORSEA: 116,
-    Pokemon.SEADRA: 117,
-    Pokemon.GOLDEEN: 118,
-    Pokemon.SEAKING: 119,
-    Pokemon.STARYU: 120,
-    Pokemon.STARMIE: 121,
-    Pokemon.MR_MIME: 122,
-    Pokemon.SCYTHER: 123,
-    Pokemon.JYNX: 124,
-    Pokemon.ELECTABUZZ: 125,
-    Pokemon.MAGMAR: 126,
-    Pokemon.PINSIR: 127,
-    Pokemon.TAUROS: 128,
-    Pokemon.MAGIKARP: 129,
-    Pokemon.GYARADOS: 130,
-    Pokemon.LAPRAS: 131,
-    Pokemon.DITTO: 132,
-    Pokemon.EEVEE: 133,
-    Pokemon.VAPOREON: 134,
-    Pokemon.JOLTEON: 135,
-    Pokemon.FLAREON: 136,
-    Pokemon.PORYGON: 137,
-    Pokemon.OMANYTE: 138,
-    Pokemon.OMASTAR: 139,
-    Pokemon.KABUTO: 140,
-    Pokemon.KABUTOPS: 141,
-    Pokemon.AERODACTYL: 142,
-    Pokemon.SNORLAX: 143,
-    Pokemon.ARTICUNO: 144,
-    Pokemon.ZAPDOS: 145,
-    Pokemon.MOLTRES: 146,
-    Pokemon.DRATINI: 147,
-    Pokemon.DRAGONAIR: 148,
-    Pokemon.DRAGONITE: 149,
-    Pokemon.MEWTWO: 150,
-    Pokemon.MEW: 151
-}
-
-
-POKEDEX_TO_INDEX_MAP = (
-    (112, 0x01),
-    (115, 0x02),
-    (32, 0x03),
-    (35, 0x04),
-    (21, 0x05),
-    (100, 0x06),
-    (34, 0x07),
-    (80, 0x08),
-    (2, 0x09),
-    (103, 0x0A),
-    (108, 0x0B),
-    (102, 0x0C),
-    (88, 0x0D),
-    (94, 0x0E),
-    (29, 0x0F),
-    (31, 0x10),
-    (104, 0x11),
-    (111, 0x12),
-    (131, 0x13),
-    (59, 0x14),
-    (151, 0x15),
-    (130, 0x16),
-    (90, 0x17),
-    (72, 0x18),
-    (92, 0x19),
-    (123, 0x1A),
-    (120, 0x1B),
-    (9, 0x1C),
-    (127, 0x1D),
-    (114, 0x1E),
-    (58, 0x21),
-    (95, 0x22),
-    (22, 0x23),
-    (16, 0x24),
-    (79, 0x25),
-    (64, 0x26),
-    (75, 0x27),
-    (113, 0x28),
-    (67, 0x29),
-    (122, 0x2A),
-    (106, 0x2B),
-    (107, 0x2C),
-    (24, 0x2D),
-    (47, 0x2E),
-    (54, 0x2F),
-    (96, 0x30),
-    (76, 0x31),
-    (126, 0x33),
-    (125, 0x35),
-    (82, 0x36),
-    (109, 0x37),
-    (56, 0x39),
-    (86, 0x3A),
-    (50, 0x3B),
-    (128, 0x3C),
-    (83, 0x40),
-    (48, 0x41),
-    (149, 0x42),
-    (84, 0x46),
-    (60, 0x47),
-    (124, 0x48),
-    (146, 0x49),
-    (144, 0x4A),
-    (145, 0x4B),
-    (132, 0x4C),
-    (52, 0x4D),
-    (98, 0x4E),
-    (37, 0x52),
-    (38, 0x53),
-    (25, 0x54),
-    (26, 0x55),
-    (147, 0x58),
-    (148, 0x59),
-    (140, 0x5A),
-    (141, 0x5B),
-    (116, 0x5C),
-    (117, 0x5D),
-    (27, 0x60),
-    (28, 0x61),
-    (138, 0x62),
-    (139, 0x63),
-    (39, 0x64),
-    (40, 0x65),
-    (133, 0x66),
-    (136, 0x67),
-    (135, 0x68),
-    (134, 0x69),
-    (66, 0x6A),
-    (41, 0x6B),
-    (23, 0x6C),
-    (46, 0x6D),
-    (61, 0x6E),
-    (62, 0x6F),
-    (13, 0x70),
-    (14, 0x71),
-    (15, 0x72),
-    (85, 0x74),
-    (57, 0x75),
-    (51, 0x76),
-    (49, 0x77),
-    (87, 0x78),
-    (10, 0x7B),
-    (11, 0x7C),
-    (12, 0x7D),
-    (68, 0x7E),
-    (55, 0x80),
-    (97, 0x81),
-    (42, 0x82),
-    (150, 0x83),
-    (143, 0x84),
-    (129, 0x85),
-    (89, 0x88),
-    (99, 0x8A),
-    (91, 0x8B),
-    (101, 0x8D),
-    (36, 0x8E),
-    (110, 0x8F),
-    (53, 0x90),
-    (105, 0x91),
-    (93, 0x93),
-    (63, 0x94),
-    (65, 0x95),
-    (17, 0x96),
-    (18, 0x97),
-    (121, 0x98),
-    (1, 0x99),
-    (2, 0x9A),
-    (73, 0x9B),
-    (118, 0x9D),
-    (119, 0x9E),
-    (77, 0xA3),
-    (78, 0xA4),
-    (19, 0xA5),
-    (20, 0xA6),
-    (33, 0xA7),
-    (30, 0xA8),
-    (74, 0xA9),
-    (137, 0xAA),
-    (142, 0xAB),
-    (81, 0xAD),
-    (4, 0xB0),
-    (7, 0xB1),
-    (5, 0xB2),
-    (8, 0xB3),
-    (6, 0xB4),
-    (43, 0xB9),
-    (44, 0xBA),
-    (45, 0xBB),
-    (69, 0xBC),
-    (70, 0xBD),
-    (71, 0xBE),
-)
