+++ pyboy/plugins/game_wrapper_pokemon_gen1/core/gen_1_memory_manager.pxd
