+++ pyboy/plugins/game_wrapper_pokemon_gen1/game_wrapper_pokemon_gen1.pxd
@@ -4,7 +4,7 @@
 #
 from libc.stdint cimport uint8_t
 from pyboy.plugins.base_plugin cimport PyBoyGameWrapper
-from pyboy.plugins.game_wrapper_pokemon_gen1.gen_1_memory_manager cimport Gen1MemoryManager
+from pyboy.plugins.game_wrapper_pokemon_gen1.core.gen_1_memory_manager cimport Gen1MemoryManager
 cimport cython
 
 cdef class GameWrapperPokemonGen1(PyBoyGameWrapper):
